refactor: add photo upload to product form
